Crate mithril_aggregator

Mithril aggregator The Aggregator is responsible for:

registering signers

producing aggregate multi-signatures

creating, storing & serving the certificate chain

creating, storing & serving certified snapshots

This crate is divided in two parts: a HTTP server that exposes an API to communicate with signers and a Runtime that tracks the blockchain to provide signed certificates. You can find more information on how it works reading the documentation website.

Re-exports

pub use dependency_injection::DependencyContainer;

pub use metrics::*;

Modules

database

database module. This module contains the entities definition tied with database representation with their associated queries.

dependency_injection

Dependency injection module. This module provides tools to initialize and share resources and services amongst different threads.

entities

Entities module

event_store

Event Store module This module proposes tools to send messages between processes and how to persist them in a separate database.

metrics

metrics module. This module contains the aggregator metrics service.

services

Services

Macros

unwrap_to_internal_server_error

Match the given result and do an early return with an internal server error (500) if it was an Error. Else return the unwrapped value.

Structs

AggregatorConfig

Configuration structure dedicated to the AggregatorRuntime.

AggregatorRunner

The runner responsibility is to expose a code API for the state machine. It holds services and configuration.

AggregatorRuntime

The AggregatorRuntime responsibility is to create a state machine to handle all actions required by the process of getting multi-signatures. See the documentation for more explanations about the Aggregator state machine.

CExplorerSignerRetriever

A SignersImporterRetriever fetching signers data from CExplorer.

CompressedArchiveSnapshotter

Compressed Archive Snapshotter create a compressed file.

Configuration

Aggregator configuration

DefaultConfiguration

Default configuration with all the default values for configurations.

DumbSnapshotter

Snapshotter that does nothing. It is mainly used for test purposes.

DumbUploader

Dummy uploader for test purposes.

FromRegisterSignerAdapter

Adapter to convert RegisterSignerMessage to Signer instances.

LocalUploader

LocalUploader is a file uploader working using local files

MainOpts

Mithril Aggregator Node

MithrilSignerRegisterer

Implementation of a SignerRegisterer

MultiSignerImpl

MultiSignerImpl is an implementation of the MultiSigner

SignerRegistrationRound

Represents the information needed to handle a signer registration round

SignersImporter

Tool that can import a list of signers

SingleSignatureAuthenticator

Authenticates single signatures against a signed message.

ToCertificatePendingMessageAdapter

Adapter to turn CertificatePending instances into CertificatePendingMessage.

ZstandardCompressionParameters

Zstandard specific parameters

Enums

CommandType

Identifies the type of command

ExecutionEnvironment

Different kinds of execution environments

RuntimeError

Error encountered or produced by the Runtime. This enum represents the faith of the errors produced during the state transitions.

SignerRegistrationError

Error type for signer registerer service.

SnapshotError

Snapshotter error type.

SnapshotUploaderType

Uploader needed to copy the snapshot once computed.

SnapshotterCompressionAlgorithm

Compression algorithm and parameters of the CompressedArchiveSnapshotter.

Traits

AggregatorRunnerTrait

This trait is intended to allow mocking the AggregatorRunner in tests. It exposes all the methods needed by the state machine.

ArtifactBuilder

ArtifactBuilder is trait for building an artifact

CertificatePendingStorer

Store for CertificatePending.

EpochSettingsStorer

Store and get aggregator epoch settings for given epoch.

FileUploader

FileUploader represents a file uploader interactor

MultiSigner

MultiSigner is the cryptographic engine in charge of producing multi signatures from individual signatures

SignerRecorder

Signer recorder trait

SignerRegisterer

Trait to register a signer

SignerRegistrationRoundOpener

Trait to open a signer registration round

SignersImporterPersister

Trait that define how a SignersImporter persist the retrieved signers.

SignersImporterRetriever

Trait that define how a SignersImporter retrieve the signers to import.

Snapshotter

Define the ability to create snapshots.

VerificationKeyStorer

Store and get signers verification keys for given epoch.