Module mithril_aggregator::database::provider

Aggregator related database providers

Structs

CertificateRecord

Certificate record is the representation of a stored certificate.

CertificateRecordProvider

Simple CertificateRecord provider.

CertificateStoreAdapter

Service to deal with certificate (read & write).

DeleteEpochSettingProvider

Provider to remove old data from the epoch_setting table

DeleteSignerRegistrationRecordProvider

Provider to remove old data from the signer_registration table

DeleteStakePoolProvider

Provider to remove old data from the stake_pool table

EpochSettingProvider

Simple EpochSettingRecord provider.

EpochSettingRecord

Settings for an epoch, including the protocol parameters.

EpochSettingStore

Service to deal with epoch settings (read & write).

InsertCertificateRecordProvider

Query to insert the certificate record

OpenMessage

OpenMessage

OpenMessageRepository

Open message repository

SignerRegistrationRecord

SignerRegistration record is the representation of a stored signer_registration.

SignerRegistrationRecordProvider

Simple SignerRegistrationRecord provider.

SignerRegistrationStoreAdapter

Service to deal with signer_registration (read & write).

StakePool

Stake pool as read from Chain.

StakePoolProvider

Simple StakePool provider.

StakePoolStore

Service to deal with stake pools (read & write).

UpdateEpochSettingProvider

Query to update the epoch setting

UpdateSignerRegistrationRecordProvider

Query to update the signer_registration record

UpdateStakePoolProvider

Query to update the stake distribution