Module mithril_aggregator::entities

Structs

Beacon

Beacon represents a point in the Cardano chain at which a Mithril certificate should be produced

Certificate

Certificate represents a Mithril certificate embedding a Mithril STM multisignature

CertificatePending

CertificatePending represents a pending certificate in the process of production

Config

Aggregator configuration

Error

Internal error representation

ProtocolParameters

Protocol cryptographic parameters

Signer

Signer represents a signing participant in the network

SignerWithStake

Signer represents a signing party in the network (including its stakes)

SingleSignature

SingleSignature represents a single signature originating from a participant in the network for a digest at a specific lottery index

Snapshot

Snapshot represents a snapshot file and its metadata

Stake

Stake represents the stakes of a participant in the Cardano chain