mithril_aggregator::services

Trait CertificateChainSynchronizer

pub trait CertificateChainSynchronizer: Send + Sync {
    // Required method
    fn synchronize_certificate_chain<'life0, 'async_trait>(
        &'life0 self,
        force: bool,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Define how to synchronize the certificate chain with a remote source

Required Methods

fn synchronize_certificate_chain<'life0, 'async_trait>( &'life0 self, force: bool, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Synchronize the certificate chain with a remote source

If force is true, the chain will always be synchronized, else it will only synchronize if the remote source has started a new chain with a new Genesis.