Trait mithril_aggregator::services::SignedEntityService

pub trait SignedEntityService: Send + Sync {
    // Required methods
    fn create_artifact<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_type: SignedEntityType,
        certificate: &'life1 Certificate,
    ) -> Pin<Box<dyn Future<Output = StdResult<JoinHandle<StdResult<()>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_last_signed_snapshots<'life0, 'async_trait>(
        &'life0 self,
        total: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntity<Snapshot>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_last_signed_mithril_stake_distributions<'life0, 'async_trait>(
        &'life0 self,
        total: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntity<MithrilStakeDistribution>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_last_cardano_transaction_snapshot<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<SignedEntity<CardanoTransactionsSnapshot>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_signed_snapshot_by_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<SignedEntity<Snapshot>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_signed_mithril_stake_distribution_by_id<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<SignedEntity<MithrilStakeDistribution>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_last_signed_cardano_stake_distributions<'life0, 'async_trait>(
        &'life0 self,
        total: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntity<CardanoStakeDistribution>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

ArtifactBuilder Service trait

Required Methods

fn create_artifact<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_type: SignedEntityType, certificate: &'life1 Certificate, ) -> Pin<Box<dyn Future<Output = StdResult<JoinHandle<StdResult<()>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create artifact for a signed entity type and a certificate

fn get_last_signed_snapshots<'life0, 'async_trait>( &'life0 self, total: usize, ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntity<Snapshot>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a list of signed snapshots order by creation date descending.

fn get_last_signed_mithril_stake_distributions<'life0, 'async_trait>( &'life0 self, total: usize, ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntity<MithrilStakeDistribution>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a list of signed Mithril stake distribution order by creation date descending.

fn get_last_cardano_transaction_snapshot<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<Option<SignedEntity<CardanoTransactionsSnapshot>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the last signed Cardano Transaction Snapshot.

fn get_signed_snapshot_by_id<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<SignedEntity<Snapshot>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return a signed snapshot

fn get_signed_mithril_stake_distribution_by_id<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<SignedEntity<MithrilStakeDistribution>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return a signed Mithril stake distribution

fn get_last_signed_cardano_stake_distributions<'life0, 'async_trait>( &'life0 self, total: usize, ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntity<CardanoStakeDistribution>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a list of signed Cardano stake distribution order by creation date descending.