mithril_aggregator::services

Trait Snapshotter

pub trait Snapshotter: Sync + Send {
    // Required methods
    fn snapshot_all_completed_immutables<'life0, 'life1, 'async_trait>(
        &'life0 self,
        archive_name_without_extension: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<FileArchive>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn snapshot_ancillary<'life0, 'life1, 'async_trait>(
        &'life0 self,
        immutable_file_number: ImmutableFileNumber,
        archive_name_without_extension: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<FileArchive>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn snapshot_immutable_trio<'life0, 'life1, 'async_trait>(
        &'life0 self,
        immutable_file_number: ImmutableFileNumber,
        archive_name_without_extension: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<FileArchive>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn compute_immutable_files_total_uncompressed_size<'life0, 'async_trait>(
        &'life0 self,
        up_to_immutable_file_number: ImmutableFileNumber,
    ) -> Pin<Box<dyn Future<Output = StdResult<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn compression_algorithm(&self) -> CompressionAlgorithm;
}

Define the ability to create snapshots.

Required Methods

fn snapshot_all_completed_immutables<'life0, 'life1, 'async_trait>( &'life0 self, archive_name_without_extension: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<FileArchive>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create a new snapshot containing all completed immutables.

fn snapshot_ancillary<'life0, 'life1, 'async_trait>( &'life0 self, immutable_file_number: ImmutableFileNumber, archive_name_without_extension: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<FileArchive>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create a new snapshot of ancillary files.

Ancillary files include the last, uncompleted, immutable trio and the two last ledger files.

fn snapshot_immutable_trio<'life0, 'life1, 'async_trait>( &'life0 self, immutable_file_number: ImmutableFileNumber, archive_name_without_extension: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<FileArchive>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create a new snapshot of an immutable trio.

fn compute_immutable_files_total_uncompressed_size<'life0, 'async_trait>( &'life0 self, up_to_immutable_file_number: ImmutableFileNumber, ) -> Pin<Box<dyn Future<Output = StdResult<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Compute the total and average uncompressed size of all immutables up to the given immutable file number.

fn compression_algorithm(&self) -> CompressionAlgorithm

Return the compression algorithm used by the snapshotter.