Crate mithril_aggregator

Structs

AggregatorConfig

AggregatorRunner

AggregatorRuntime

AggregatorRuntime

BeaconProviderImpl

CertificatePendingStore

CertificateStore

Config

Aggregator configuration

DependencyManager

DependencyManager handles the dependencies

DumbImmutableFileObserver

DumbSnapshotUploader

DumbSnapshotter

GzipSnapshotter

Snapshotter

ImmutableFileSystemObserver

LocalSnapshotStore

LocalSnapshotUploader

LocalSnapshotUploader is a snapshot uploader working using local files

MemoryBeaconStore

MemoryBeaconStore is in memory BeaconStore

MultiSignerImpl

MultiSignerImpl is an implementation of the MultiSigner

RemoteSnapshotStore

GoogleCloudPlatformSnapshotStore is a snapshot store working using Google Cloud Platform services

RemoteSnapshotUploader

GCPSnapshotUploader is a snapshot uploader working using Google Cloud Platform services

Server

Server

SingleSignatureStore

VerificationKeyStore

Enums

BeaconStoreError

ProtocolError

SnapshotError

SnapshotStoreType

SnapshotUploaderType

VerificationKeyStoreError

Traits

AggregatorRunnerTrait

BeaconProvider

BeaconStore

BeaconStore represents a beacon store interactor

ImmutableFileObserver

MultiSigner

MultiSigner is the cryptographic engine in charge of producing multi signatures from individual signatures

SnapshotStore

SnapshotStore represents a snapshot store interactor

SnapshotUploader

SnapshotUploader represents a snapshot uploader interactor

Snapshotter

VerificationKeyStorer