Crate mithril_aggregator

Structs

AggregatorRuntime

AggregatorRuntime

Config

Aggregator configuration

DependencyManager

DependencyManager handles the dependencies

LocalSnapshotUploader

LocalSnapshotUploader is a snapshot uploader working using local files

MemoryBeaconStore

MemoryBeaconStore is in memory [BeaconStore]

MultiSignerImpl

MultiSignerImpl is an implementation of the MultiSigner

RemoteSnapshotStore

GoogleCloudPlatformSnapshotStore is a snapshot store working using Google Cloud Platform services

RemoteSnapshotUploader

GCPSnapshotUploader is a snapshot uploader working using Google Cloud Platform services

Server

Server

Snapshotter

Snapshotter

Enums

BeaconStoreError

ProtocolError

SnapshotError

Traits

MultiSigner

MultiSigner is the cryptographic engine in charge of producing multi signatures from individual signatures

SnapshotStore

SnapshotStore represents a snapshot store interactor