Crate mithril_aggregator

Mithril aggregator The Aggregator is responsible for:

registering signers

producing aggregate multi-signatures

creating, storing & serving the certificate chain

creating, storing & serving certified snapshots

This crate is divided in two parts: a HTTP server that exposes an API to communicate with signers and a Runtime that tracks the blockchain to provide signed certificates. You can find more information on how it works reading the documentation website.

Structs

AggregatorConfig

Configuration structure dedicated to the AggregatorRuntime.

AggregatorRunner

The runner responsibility is to expose a code API for the state machine. It holds services and configuration.

AggregatorRuntime

The AggregatorRuntime responsibility is to create a state machine to handle all actions required by the process of getting multi-signatures. See the documentation for more explanations about the Aggregator state machine.

CertificatePendingStore

Store for [CertificatePending].

CertificateStore

Store for issued certificates.

Configuration

Aggregator configuration

DefaultConfiguration

Default configuration with all the default values for configurations.

DependencyManager

DependencyManager handles the dependencies

DumbSnapshotUploader

Dummy uploader for test purposes.

DumbSnapshotter

Snapshotter that does nothing. It is mainly used for test purposes.

FromRegisterSignerAdapter

Adapter to convert [RegisterSignerMessage] to [Signer] instances.

GenesisConfiguration

Configuration expected for Genesis commands.

GzipSnapshotter

Gzip Snapshotter create a compressed file.

LocalSnapshotStore

Store for local snapshots.

LocalSnapshotUploader

LocalSnapshotUploader is a snapshot uploader working using local files

MainOpts

Mithril Aggregator Node

MithrilCertificateCreator

Implementation of a CertificateCreator

MithrilSignerRegisterer

Implementation of a SignerRegisterer

MultiSignerImpl

MultiSignerImpl is an implementation of the MultiSigner

ProtocolParametersStore

ProtocolParameter store.

RemoteSnapshotStore

GoogleCloudPlatformSnapshotStore is a snapshot store working using Google Cloud Platform services

RemoteSnapshotUploader

GCPSnapshotUploader is a snapshot uploader working using Google Cloud Platform services

Server

Server

SignerRegistrationRound

Represents the information needed to handle a signer registration round

SingleSignatureStore

Store for [SingleSignatures].

ToCertificatePendingMessageAdapter

Adapter to turn [CertificatePending] instances into [CertificatePendingMessage].

ToEpochSettingsMessageAdapter

Adapter to spawn [EpochSettingsMessage] from [EpochSettings] instances.

VerificationKeyStore

Store for the VerificationKey.

Enums

ProtocolError

Error type for multi signer service.

SignerRegistrationError

Error type for signer registerer service.

SnapshotError

Snapshotter error type.

SnapshotStoreType

Snapshot store type enumerates the different kinds of snapshot stores. Local storage is mainly used by development and test environements while GCP is intended for production use.

SnapshotUploaderType

Uploader needed to copy the snapshot once computed.

Traits

AggregatorRunnerTrait

This trait is intended to allow mocking the AggregatorRunner in tests. It exposes all the methods needed by the state machine.

CertificateCreator

Define a way to create a [Certificate][mithril_common::entities::Certificate]

MultiSigner

MultiSigner is the cryptographic engine in charge of producing multi signatures from individual signatures

ProtocolParametersStorer

Trait for mocking ProtocolParameterStore.

SignerRegisterer

Trait to register a signer

SignerRegistrationRoundOpener

Trait to open a signer registration round

SnapshotStore

SnapshotStore represents a snapshot store interactor

SnapshotUploader

SnapshotUploader represents a snapshot uploader interactor

Snapshotter

Define the ability to create snapshots.

VerificationKeyStorer

Mocking trait for VerificationKeyStore.