mithril_aggregator::database

Module record

Aggregator related database records

Structs

BufferedSingleSignatureRecord

BufferedSingleSignatureRecord record is the representation of a buffered single_signature that may be used for upcoming open messages.

CertificateRecord

Certificate record is the representation of a stored certificate.

EpochSettingsRecord

Settings for an epoch, including the protocol parameters.

ImmutableFileDigestRecord

ImmutableFileDigestRecord is the record that stores the digest of an immutable file.

OpenMessageRecord

OpenMessage

OpenMessageWithSingleSignaturesRecord

Open Message with associated single signatures if any.

SignedEntityRecord

SignedEntity record is the representation of a stored signed_entity.

SignerRecord

Signer record is the representation of a stored signer.

SignerRegistrationRecord

SignerRegistration record is the representation of a stored signer_registration.

SingleSignatureRecord

SingleSignature record is the representation of a stored single_signature.

StakePool

Stake pool as read from Chain.