mithril_aggregator

Module services

Services

This module regroups services. Services are adapters in charge of the different bounded contexts of the application:

Ticker: provides the time of the blockchain

StakeEntity: fetches Cardano stake distribution information

Certifier: registers signers and create certificates once ready

SignedEntity: provides information about signed entities.

Each service is defined by a public API (a trait) that is used in the controllers (runtimes).

Modules

ancillary_signer

Tools for signing generating signed manifests of ancillary data.

Structs

AggregatorChainDataImporter

Services responsible for importing chain data into the aggregator data stores.

AggregatorSignableSeedBuilder

SignableSeedBuilder aggregator implementation

AggregatorUpkeepService

Implementation of the upkeep service for the aggregator.

BufferedCertifierService

A decorator of CertifierService that can buffer registration of single signatures when the open message is not yet created.

CompressedArchiveSnapshotter

Compressed Archive Snapshotter create a compressed file.

DumbSnapshotter

Snapshotter that does nothing. It is mainly used for test purposes.

EpochServiceDependencies

Dependencies required by the MithrilEpochService.

FakeSignatureConsumer

A fake implementation of the SignatureConsumer trait (test only).

FakeSnapshotter

Snapshotter that writes empty files to the filesystem. Used for testing purposes.

LegacyMithrilProverService

Legacy Mithril prover

LocalMithrilNetworkConfigurationProvider

Read network configuration from the database epoch_settings or, if no records is available for an epoch, fallback to the provided configuration value

MithrilCertificateChainSynchronizer

Service that synchronizes the certificate chain with a remote aggregator

MithrilCertificateChainSynchronizerNoop

A noop CertificateChainSynchronizer for leader aggregators

MithrilCertifierService

Mithril CertifierService implementation

MithrilEpochService

Implementation of the epoch service.

MithrilMessageService

Implementation of the MessageService

MithrilProverService

Mithril prover

MithrilSignedEntityService

Mithril ArtifactBuilder Service

MithrilSignerRegistrationFollower

A MithrilSignerRegistrationFollower supports signer registrations in a follower aggregator

MithrilSignerRegistrationLeader

A MithrilSignerRegistrationLeader supports signer registrations in a leader aggregator

MithrilSignerRegistrationVerifier

Implementation of a SignerRegistrationVerifier

MithrilStakeDistributionService

Implementation of the stake distribution service.

SequentialSignatureProcessor

A sequential signature processor receives messages and processes them sequentially

SignatureConsumerDmq

DMQ implementation of the SignatureConsumer trait.

SignatureConsumerNoop

A no-op implementation of the SignatureConsumer trait that will never return signatures.

SignedEntityServiceArtifactsDependencies

ArtifactsBuilder dependencies required by the MithrilSignedEntityService.

SignerRegistrationRound

Represents the information needed to handle a signer registration round

UsageReporter

Reporter of usage metrics of the application.

Enums

CertifierServiceError

Errors dedicated to the CertifierService.

EpochServiceError

Errors dedicated to the CertifierService.

SignatureRegistrationStatus

Status of a successful registration of a single signature.

SignerRegistrationError

Error type for signer registerer service.

StakePoolDistributionServiceError

Errors related to the StakeDistributionService.

Traits

BufferedSingleSignatureStore

BufferedSingleSignatureStore

CertificateChainSynchronizer

Define how to synchronize the certificate chain with a remote source

CertifierService

CertifierService

EpochPruningTask

Define the task responsible for pruning a datasource below a certain epoch threshold.

EpochService

Service that aggregates all data that don’t change in a given epoch.

LeaderAggregatorClient

Define how data are retrieved from a leader aggregator

LegacyProverService

Legacy Prover service is the cryptographic engine in charge of producing cryptographic proofs for transactions

MessageService

HTTP Message service trait.

OpenMessageStorer

Define how to store the open message created at the end of the synchronization process

ProverService

Prover service is the cryptographic engine in charge of producing cryptographic proofs for transactions and blocks

RemoteCertificateRetriever

Define how to retrieve remote certificate details

SignatureConsumer

A signature consumer which blocks until messages are available.

SignatureProcessor

A signature processor which receives signature and processes them.

SignedEntityService

ArtifactBuilder Service trait

SignerRecorder

Signer recorder trait

SignerRegisterer

Trait to register a signer

SignerRegistrationRoundOpener

Trait to open a signer registration round

SignerRegistrationVerifier

A trait for verifying a Signer registration.

SignerSynchronizer

Trait to synchronize signers

Snapshotter

Define the ability to create snapshots.

StakeDistributionService

Responsible of synchronizing with Cardano stake distribution.

SynchronizedCertificateStorer

Define how to store the synchronized certificate and retrieve details about the actual local chain

TransactionsRetriever

Transactions retriever

UpkeepService

Define the service responsible for the upkeep of the application.