Module mithril_aggregator::services

Services

This module regroups services. Services are adapters in charge of the different bounded contexts of the application:

Ticker: provides the time of the blockchain

StakeEntity: fetches Cardano stake distribution information

Certifier: registers signers and create certificates once ready

SignedEntity: provides information about signed entities.

Each service is defined by a public API (a trait) that is used in the controllers (runtimes).

Structs

CardanoTransactionsImporter

Import and store CardanoTransaction.

MithrilCertifierService

Mithril CertifierService implementation

MithrilEpochService

Implementation of the epoch service.

MithrilMessageService

Implementation of the MessageService

MithrilProverService

Mithril prover

MithrilSignedEntityService

Mithril ArtifactBuilder Service

MithrilStakeDistributionService

Implementation of the stake distribution service.

MithrilTickerService

MithrilTickerService

Enums

CertifierServiceError

Errors dedicated to the CertifierService.

EpochServiceError

Errors dedicated to the CertifierService.

MessageServiceError

Error related to the MessageService

StakePoolDistributionServiceError

Errors related to the StakeDistributionService.

Traits

CertifierService

CertifierService

EpochService

Service that aggregates all data that don’t change in a given epoch.

MessageService

HTTP Message service trait.

ProverService

Prover service is the cryptographic engine in charge of producing cryptographic proofs for transactions

SignedEntityService

ArtifactBuilder Service trait

StakeDistributionService

Responsible of synchronizing with Cardano stake distribution.

TickerService

Service trait with consistent business oriented API.

TransactionStore

Cardano transactions store

TransactionsRetriever

Transactions retriever