mithril_aggregator::services

Trait MessageService

pub trait MessageService: Sync + Send {

    // Required methods
    fn get_epoch_settings_message<'life0, 'async_trait>(
        &'life0 self,
        allowed_discriminants: BTreeSet<SignedEntityTypeDiscriminants>,
    ) -> Pin<Box<dyn Future<Output = StdResult<EpochSettingsMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_protocol_configuration_message<'life0, 'async_trait>(
        &'life0 self,
        epoch: Epoch,
        allowed_discriminants: BTreeSet<SignedEntityTypeDiscriminants>,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<ProtocolConfigurationMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_certificate_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate_hash: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CertificateMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_latest_genesis_certificate_message<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CertificateMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_certificate_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<CertificateListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_snapshot_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<SnapshotMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_snapshot_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<SnapshotListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_database_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoDatabaseSnapshotMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_cardano_database_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<CardanoDatabaseSnapshotListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_database_list_message_by_epoch<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
        epoch: Epoch,
    ) -> Pin<Box<dyn Future<Output = StdResult<CardanoDatabaseSnapshotListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_database_digest_list_message<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = StdResult<CardanoDatabaseDigestListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_mithril_stake_distribution_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<MithrilStakeDistributionMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_mithril_stake_distribution_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<MithrilStakeDistributionListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_transaction_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoTransactionSnapshotMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_cardano_transaction_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<CardanoTransactionSnapshotListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_blocks_transactions_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoBlocksTransactionsSnapshotMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_cardano_blocks_transactions_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<CardanoBlocksTransactionsSnapshotListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_stake_distribution_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_id: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoStakeDistributionMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_cardano_stake_distribution_message_by_epoch<'life0, 'async_trait>(
        &'life0 self,
        epoch: Epoch,
    ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoStakeDistributionMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_cardano_stake_distribution_list_message<'life0, 'async_trait>(
        &'life0 self,
        limit: usize,
    ) -> Pin<Box<dyn Future<Output = StdResult<CardanoStakeDistributionListMessage>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

HTTP Message service trait.

Required Methods

fn get_epoch_settings_message<'life0, 'async_trait>( &'life0 self, allowed_discriminants: BTreeSet<SignedEntityTypeDiscriminants>, ) -> Pin<Box<dyn Future<Output = StdResult<EpochSettingsMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the epoch settings message if it exists.

fn get_protocol_configuration_message<'life0, 'async_trait>( &'life0 self, epoch: Epoch, allowed_discriminants: BTreeSet<SignedEntityTypeDiscriminants>, ) -> Pin<Box<dyn Future<Output = StdResult<Option<ProtocolConfigurationMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the protocol configuration message for the given epoch if it exists.

fn get_certificate_message<'life0, 'life1, 'async_trait>( &'life0 self, certificate_hash: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CertificateMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the message representation of a certificate if it exists.

fn get_latest_genesis_certificate_message<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CertificateMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the message representation of the latest genesis certificate.

fn get_certificate_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<CertificateListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the message representation of the last N certificates.

fn get_snapshot_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<SnapshotMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the information regarding the given snapshot.

fn get_snapshot_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<SnapshotListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last signed snapshots. The limit of the list is passed as argument.

fn get_cardano_database_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoDatabaseSnapshotMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the information regarding the Cardano database for the given identifier.

fn get_cardano_database_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<CardanoDatabaseSnapshotListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last Cardano database message.

fn get_cardano_database_list_message_by_epoch<'life0, 'async_trait>( &'life0 self, limit: usize, epoch: Epoch, ) -> Pin<Box<dyn Future<Output = StdResult<CardanoDatabaseSnapshotListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last Cardano database message.

fn get_cardano_database_digest_list_message<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = StdResult<CardanoDatabaseDigestListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the Cardano database immutable file names and their digests.

fn get_mithril_stake_distribution_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<MithrilStakeDistributionMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the information regarding the Mithril stake distribution for the given identifier.

fn get_mithril_stake_distribution_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<MithrilStakeDistributionListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last Mithril stake distributions message.

fn get_cardano_transaction_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoTransactionSnapshotMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the information regarding the Cardano transactions set for the given identifier.

fn get_cardano_transaction_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<CardanoTransactionSnapshotListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last Cardano transactions set message.

fn get_cardano_blocks_transactions_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoBlocksTransactionsSnapshotMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the information regarding the Cardano blocks transactions set for the given identifier.

fn get_cardano_blocks_transactions_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<CardanoBlocksTransactionsSnapshotListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last Cardano blocks transactions set message.

fn get_cardano_stake_distribution_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_id: &'life1 str, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoStakeDistributionMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the information regarding the Cardano stake distribution for the given identifier.

fn get_cardano_stake_distribution_message_by_epoch<'life0, 'async_trait>( &'life0 self, epoch: Epoch, ) -> Pin<Box<dyn Future<Output = StdResult<Option<CardanoStakeDistributionMessage>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the information regarding the Cardano stake distribution for the given epoch.

fn get_cardano_stake_distribution_list_message<'life0, 'async_trait>( &'life0 self, limit: usize, ) -> Pin<Box<dyn Future<Output = StdResult<CardanoStakeDistributionListMessage>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the list of the last Cardano stake distributions message.