Trait mithril_aggregator::AggregatorRunnerTrait

pub trait AggregatorRunnerTrait: Sync + Send {

    fn get_beacon_from_chain<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Beacon, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn does_certificate_exist_for_beacon<'life0, 'life1, 'async_trait>(
        &'life0 self,
        beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<bool, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn is_certificate_chain_valid<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<bool, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn compute_digest<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<String, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn update_beacon<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn update_stake_distribution<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn update_protocol_parameters_in_multisigner<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn update_message_in_multisigner<'life0, 'async_trait>(
        &'life0 self,
        digest: String
    ) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn create_new_pending_certificate_from_multisigner<'life0, 'async_trait>(
        &'life0 self,
        beacon: Beacon
    ) -> Pin<Box<dyn Future<Output = Result<CertificatePending, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn create_new_working_certificate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate_pending: &'life1 CertificatePending
    ) -> Pin<Box<dyn Future<Output = Result<WorkingCertificate, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn save_pending_certificate<'life0, 'async_trait>(
        &'life0 self,
        pending_certificate: CertificatePending
    ) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn drop_pending_certificate<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<CertificatePending>, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn is_multisig_created<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<bool, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn create_snapshot_archive<'life0, 'life1, 'async_trait>(
        &'life0 self,
        beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<OngoingSnapshot, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn upload_snapshot_archive<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ongoing_snapshot: &'life1 OngoingSnapshot
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn create_and_save_certificate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        working_certificate: &'life1 WorkingCertificate
    ) -> Pin<Box<dyn Future<Output = Result<Certificate, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn create_and_save_snapshot<'life0, 'life1, 'async_trait>(
        &'life0 self,
        certificate: Certificate,
        ongoing_snapshot: &'life1 OngoingSnapshot,
        remote_locations: Vec<String>
    ) -> Pin<Box<dyn Future<Output = Result<Snapshot, RuntimeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

}

This trait is intended to allow mocking the AggregatorRunner in tests. It exposes all the methods needed by the state machine.

Required Methods

fn get_beacon_from_chain<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Beacon, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Return the current beacon from the chain

fn does_certificate_exist_for_beacon<'life0, 'life1, 'async_trait>(
&'life0 self,
beacon: &'life1 Beacon
) -> Pin<Box<dyn Future<Output = Result<bool, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Check if a certificate already have been issued for a given beacon.

fn is_certificate_chain_valid<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<bool, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Check if a certificate chain is valid.

fn compute_digest<'life0, 'life1, 'async_trait>(
&'life0 self,
new_beacon: &'life1 Beacon
) -> Pin<Box<dyn Future<Output = Result<String, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Compute the digest of the last immutable file of the node.

fn update_beacon<'life0, 'life1, 'async_trait>(
&'life0 self,
new_beacon: &'life1 Beacon
) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Update the multisigner with the given beacon.

fn update_stake_distribution<'life0, 'life1, 'async_trait>(
&'life0 self,
new_beacon: &'life1 Beacon
) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Read the stake distribution from the blockchain and store it.

fn update_protocol_parameters_in_multisigner<'life0, 'life1, 'async_trait>(
&'life0 self,
new_beacon: &'life1 Beacon
) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Update the multisigner with the protocol parameters from configuration.

fn update_message_in_multisigner<'life0, 'async_trait>(
&'life0 self,
digest: String
) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Set the message to sign in the multisigner. The digest is only one part of the message, the next signing stake distribution must also be signed as part of the message.

fn create_new_pending_certificate_from_multisigner<'life0, 'async_trait>(
&'life0 self,
beacon: Beacon
) -> Pin<Box<dyn Future<Output = Result<CertificatePending, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Return the actual pending certificate from the multisigner.

fn create_new_working_certificate<'life0, 'life1, 'async_trait>(
&'life0 self,
certificate_pending: &'life1 CertificatePending
) -> Pin<Box<dyn Future<Output = Result<WorkingCertificate, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Return the actual working certificate from the multisigner.

fn save_pending_certificate<'life0, 'async_trait>(
&'life0 self,
pending_certificate: CertificatePending
) -> Pin<Box<dyn Future<Output = Result<(), RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Store the given pending certificate.

fn drop_pending_certificate<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<CertificatePending>, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Drop the multisigner’s actual pending certificate.

fn is_multisig_created<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<bool, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Check if the multisigner has issued a multi-signature.

fn create_snapshot_archive<'life0, 'life1, 'async_trait>(
&'life0 self,
beacon: &'life1 Beacon
) -> Pin<Box<dyn Future<Output = Result<OngoingSnapshot, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Create an archive of the cardano node db directory naming it after the given beacon.

Returns the path of the created archive and the archive size as byte.

fn upload_snapshot_archive<'life0, 'life1, 'async_trait>(
&'life0 self,
ongoing_snapshot: &'life1 OngoingSnapshot
) -> Pin<Box<dyn Future<Output = Result<Vec<String>, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Upload the snapshot at the given location using the configured uploader(s).

Important: the snapshot is removed after the upload succeeded.

fn create_and_save_certificate<'life0, 'life1, 'async_trait>(
&'life0 self,
working_certificate: &'life1 WorkingCertificate
) -> Pin<Box<dyn Future<Output = Result<Certificate, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Create a signed certificate.

fn create_and_save_snapshot<'life0, 'life1, 'async_trait>(
&'life0 self,
certificate: Certificate,
ongoing_snapshot: &'life1 OngoingSnapshot,
remote_locations: Vec<String>
) -> Pin<Box<dyn Future<Output = Result<Snapshot, RuntimeError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Create a snapshot and save it to the given locations.