Trait mithril_aggregator::AggregatorRunnerTrait

pub trait AggregatorRunnerTrait: Sync + Send {

    // Required methods
    fn get_beacon_from_chain<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Beacon, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_current_non_certified_open_message_for_signed_entity_type<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_type: &'life1 SignedEntityType
    ) -> Pin<Box<dyn Future<Output = Result<Option<OpenMessage>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_current_non_certified_open_message<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<OpenMessage>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn is_certificate_chain_valid<'life0, 'life1, 'async_trait>(
        &'life0 self,
        beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_beacon<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn update_stake_distribution<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn open_signer_registration_round<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn close_signer_registration_round<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_protocol_parameters_in_multisigner<'life0, 'life1, 'async_trait>(
        &'life0 self,
        new_beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn compute_protocol_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_type: &'life1 SignedEntityType
    ) -> Pin<Box<dyn Future<Output = Result<ProtocolMessage, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn create_new_pending_certificate_from_multisigner<'life0, 'life1, 'async_trait>(
        &'life0 self,
        beacon: Beacon,
        signed_entity_type: &'life1 SignedEntityType
    ) -> Pin<Box<dyn Future<Output = Result<CertificatePending, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn save_pending_certificate<'life0, 'async_trait>(
        &'life0 self,
        pending_certificate: CertificatePending
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop_pending_certificate<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<CertificatePending>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_certificate<'life0, 'life1, 'async_trait>(
        &'life0 self,
        signed_entity_type: &'life1 SignedEntityType
    ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn create_artifact<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        signed_entity_type: &'life1 SignedEntityType,
        certificate: &'life2 Certificate
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn update_era_checker<'life0, 'life1, 'async_trait>(
        &'life0 self,
        beacon: &'life1 Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn certifier_inform_new_epoch<'life0, 'life1, 'async_trait>(
        &'life0 self,
        epoch: &'life1 Epoch
    ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn create_open_message<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        signed_entity_type: &'life1 SignedEntityType,
        protocol_message: &'life2 ProtocolMessage
    ) -> Pin<Box<dyn Future<Output = Result<OpenMessage, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

}

This trait is intended to allow mocking the AggregatorRunner in tests. It exposes all the methods needed by the state machine.

Required Methods

fn get_beacon_from_chain<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Beacon, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Return the current beacon from the chain

fn get_current_non_certified_open_message_for_signed_entity_type<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_type: &'life1 SignedEntityType ) -> Pin<Box<dyn Future<Output = Result<Option<OpenMessage>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieves the current non certified open message for a given signed entity type.

fn get_current_non_certified_open_message<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<OpenMessage>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the current non certified open message.

fn is_certificate_chain_valid<'life0, 'life1, 'async_trait>( &'life0 self, beacon: &'life1 Beacon ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Check if a certificate chain is valid.

fn update_beacon<'life0, 'life1, 'async_trait>( &'life0 self, new_beacon: &'life1 Beacon ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Update the multisigner with the given beacon.

fn update_stake_distribution<'life0, 'life1, 'async_trait>( &'life0 self, new_beacon: &'life1 Beacon ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Read the stake distribution from the blockchain and store it.

fn open_signer_registration_round<'life0, 'life1, 'async_trait>( &'life0 self, new_beacon: &'life1 Beacon ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Open the signer registration round of an epoch.

fn close_signer_registration_round<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Close the signer registration round of an epoch.

fn update_protocol_parameters_in_multisigner<'life0, 'life1, 'async_trait>( &'life0 self, new_beacon: &'life1 Beacon ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Update the multisigner with the protocol parameters from configuration.

fn compute_protocol_message<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_type: &'life1 SignedEntityType ) -> Pin<Box<dyn Future<Output = Result<ProtocolMessage, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Compute the protocol message

fn create_new_pending_certificate_from_multisigner<'life0, 'life1, 'async_trait>( &'life0 self, beacon: Beacon, signed_entity_type: &'life1 SignedEntityType ) -> Pin<Box<dyn Future<Output = Result<CertificatePending, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Return the actual pending certificate from the multisigner.

fn save_pending_certificate<'life0, 'async_trait>( &'life0 self, pending_certificate: CertificatePending ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Store the given pending certificate.

fn drop_pending_certificate<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<CertificatePending>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Drop the multisigner’s actual pending certificate.

fn create_certificate<'life0, 'life1, 'async_trait>( &'life0 self, signed_entity_type: &'life1 SignedEntityType ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Create multi-signature.

fn create_artifact<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, signed_entity_type: &'life1 SignedEntityType, certificate: &'life2 Certificate ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Create an artifact and persist it.

fn update_era_checker<'life0, 'life1, 'async_trait>( &'life0 self, beacon: &'life1 Beacon ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Update the EraChecker with EraReader information.

fn certifier_inform_new_epoch<'life0, 'life1, 'async_trait>( &'life0 self, epoch: &'life1 Epoch ) -> Pin<Box<dyn Future<Output = Result<(), Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Certifier inform new epoch

fn create_open_message<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, signed_entity_type: &'life1 SignedEntityType, protocol_message: &'life2 ProtocolMessage ) -> Pin<Box<dyn Future<Output = Result<OpenMessage, Box<dyn StdError + Sync + Send>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Create new open message