mithril_aggregator

Trait ConfigurationSource

pub trait ConfigurationSource {

    // Required method
    fn environment(&self) -> ExecutionEnvironment;

    // Provided methods
    fn cardano_cli_path(&self) -> PathBuf { ... }
    fn cardano_node_socket_path(&self) -> PathBuf { ... }
    fn dmq_node_socket_path(&self) -> Option<PathBuf> { ... }
    fn cardano_node_version(&self) -> String { ... }
    fn network(&self) -> String { ... }
    fn network_magic(&self) -> Option<u64> { ... }
    fn dmq_network_magic(&self) -> Option<u64> { ... }
    fn chain_observer_type(&self) -> ChainObserverType { ... }
    fn protocol_parameters(&self) -> Option<ProtocolParameters> { ... }
    fn snapshot_uploader_type(&self) -> SnapshotUploaderType { ... }
    fn snapshot_bucket_name(&self) -> Option<String> { ... }
    fn snapshot_use_cdn_domain(&self) -> bool { ... }
    fn server_ip(&self) -> String { ... }
    fn server_port(&self) -> u16 { ... }
    fn public_server_url(&self) -> Option<String> { ... }
    fn run_interval(&self) -> u64 { ... }
    fn db_directory(&self) -> PathBuf { ... }
    fn snapshot_directory(&self) -> PathBuf { ... }
    fn data_stores_directory(&self) -> PathBuf { ... }
    fn genesis_verification_key(&self) -> HexEncodedGenesisVerificationKey { ... }
    fn reset_digests_cache(&self) -> bool { ... }
    fn disable_digests_cache(&self) -> bool { ... }
    fn store_retention_limit(&self) -> Option<usize> { ... }
    fn era_reader_adapter_type(&self) -> EraReaderAdapterType { ... }
    fn era_reader_adapter_params(&self) -> Option<String> { ... }
    fn ancillary_files_signer_config(&self) -> AncillaryFilesSignerConfig { ... }
    fn signed_entity_types(&self) -> Option<String> { ... }
    fn snapshot_compression_algorithm(&self) -> CompressionAlgorithm { ... }
    fn zstandard_parameters(&self) -> Option<ZstandardCompressionParameters> { ... }
    fn blockfrost_parameters(&self) -> Option<BlockfrostParameters> { ... }
    fn signer_importer_run_interval(&self) -> u64 { ... }
    fn allow_unparsable_block(&self) -> bool { ... }
    fn cardano_blocks_transactions_prover_cache_pool_size(&self) -> usize { ... }
    fn cardano_blocks_transactions_database_connection_pool_size(&self) -> usize { ... }
    fn cardano_transactions_prover_cache_pool_size(&self) -> usize { ... }
    fn cardano_transactions_database_connection_pool_size(&self) -> usize { ... }
    fn cardano_transactions_signing_config(
        &self,
    ) -> Option<CardanoTransactionsSigningConfig> { ... }
    fn cardano_blocks_transactions_signing_config(
        &self,
    ) -> Option<CardanoBlocksTransactionsSigningConfig> { ... }
    fn preload_security_parameter(&self) -> BlockNumber { ... }
    fn cardano_prover_max_hashes_allowed_by_request(&self) -> usize { ... }
    fn cardano_transactions_block_streamer_max_roll_forwards_per_poll(
        &self,
    ) -> usize { ... }
    fn cardano_transactions_block_streamer_throttling_interval(
        &self,
    ) -> Option<u64> { ... }
    fn enable_metrics_server(&self) -> bool { ... }
    fn metrics_server_ip(&self) -> String { ... }
    fn metrics_server_port(&self) -> u16 { ... }
    fn persist_usage_report_interval_in_seconds(&self) -> u64 { ... }
    fn leader_aggregator_endpoint(&self) -> Option<String> { ... }
    fn custom_origin_tag_white_list(&self) -> Option<String> { ... }
    fn get_server_url(&self) -> StdResult<SanitizedUrlWithTrailingSlash> { ... }
    fn get_network(&self) -> StdResult<CardanoNetwork> { ... }
    fn get_dmq_network(&self) -> StdResult<DmqNetwork> { ... }
    fn get_sqlite_dir(&self) -> PathBuf { ... }
    fn get_snapshot_dir(&self) -> StdResult<PathBuf> { ... }
    fn safe_epoch_retention_limit(&self) -> Option<u64> { ... }
    fn compute_allowed_signed_entity_types_discriminants(
        &self,
    ) -> StdResult<BTreeSet<SignedEntityTypeDiscriminants>> { ... }
    fn allow_http_serve_directory(&self) -> bool { ... }
    fn get_leader_aggregator_epoch_settings_configuration(
        &self,
    ) -> StdResult<AggregatorEpochSettings> { ... }
    fn is_follower_aggregator(&self) -> bool { ... }
    fn compute_origin_tag_white_list(&self) -> HashSet<String> { ... }
    fn aggregate_signature_type(&self) -> AggregateSignatureType { ... }
    fn signature_processor_wait_delay_on_error_ms(&self) -> u64 { ... }

}

This trait defines the configuration interface for the aggregator.

By default, each function panics if not overridden, forcing concrete configuration implementations to explicitly provide the necessary values.

Required Methods

fn environment(&self) -> ExecutionEnvironment

What kind of runtime environment the configuration is meant to.

Provided Methods

fn cardano_cli_path(&self) -> PathBuf

Cardano CLI tool path

fn cardano_node_socket_path(&self) -> PathBuf

Path of the socket opened by the Cardano node

fn dmq_node_socket_path(&self) -> Option<PathBuf>

Path of the socket opened by the DMQ node

fn cardano_node_version(&self) -> String

Cardano node version.

NOTE: This cannot be verified for now (see this issue). This is why it has to be manually given to the Aggregator

fn network(&self) -> String

Cardano network

fn network_magic(&self) -> Option<u64>

Cardano Network Magic number

useful for TestNet & DevNet

fn dmq_network_magic(&self) -> Option<u64>

DMQ Network Magic number

useful for TestNet & DevNet

fn chain_observer_type(&self) -> ChainObserverType

Cardano chain observer type

fn protocol_parameters(&self) -> Option<ProtocolParameters>

Protocol parameters

fn snapshot_uploader_type(&self) -> SnapshotUploaderType

Type of snapshot uploader to use

fn snapshot_bucket_name(&self) -> Option<String>

Bucket name where the snapshots are stored if snapshot_uploader_type is Gcp

fn snapshot_use_cdn_domain(&self) -> bool

Use CDN domain to construct snapshot urls if snapshot_uploader_type is Gcp

fn server_ip(&self) -> String

Server listening IP

fn server_port(&self) -> u16

Server listening port

fn public_server_url(&self) -> Option<String>

Server URL that can be accessed from the outside

fn run_interval(&self) -> u64

Run Interval is the interval between two runtime cycles in ms

fn db_directory(&self) -> PathBuf

Directory of the Cardano node store.

fn snapshot_directory(&self) -> PathBuf

Directory to store snapshot

fn data_stores_directory(&self) -> PathBuf

Directory to store aggregator databases

fn genesis_verification_key(&self) -> HexEncodedGenesisVerificationKey

Genesis verification key

fn reset_digests_cache(&self) -> bool

Should the immutable cache be reset or not

fn disable_digests_cache(&self) -> bool

Use the digest caching strategy

fn store_retention_limit(&self) -> Option<usize>

Max number of records in stores. When new records are added, oldest records are automatically deleted so there can always be at max the number of records specified by this setting.

fn era_reader_adapter_type(&self) -> EraReaderAdapterType

Era reader adapter type

fn era_reader_adapter_params(&self) -> Option<String>

Era reader adapter parameters

fn ancillary_files_signer_config(&self) -> AncillaryFilesSignerConfig

Configuration of the ancillary files signer

IMPORTANT: The cryptographic scheme used is ED25519

fn signed_entity_types(&self) -> Option<String>

Signed entity types parameters (discriminants names in an ordered, case-sensitive, comma separated list).

The values MithrilStakeDistribution and CardanoImmutableFilesFull are prepended automatically to the list.

fn snapshot_compression_algorithm(&self) -> CompressionAlgorithm

Compression algorithm used for the snapshot archive artifacts.

fn zstandard_parameters(&self) -> Option<ZstandardCompressionParameters>

Specific parameters when CompressionAlgorithm is set to zstandard.

fn blockfrost_parameters(&self) -> Option<BlockfrostParameters>

Blockfrost api configuration

fn signer_importer_run_interval(&self) -> u64

Time interval at which the pools names and ticker in blockfrost will be imported (in minutes).

fn allow_unparsable_block(&self) -> bool

If set no error is returned in case of unparsable block and an error log is written instead.

Will be ignored on (pre)production networks.

fn cardano_blocks_transactions_prover_cache_pool_size(&self) -> usize

Cardano blocks and transactions prover cache pool size

fn cardano_blocks_transactions_database_connection_pool_size(&self) -> usize

Cardano and transactions database connection pool size

fn cardano_transactions_prover_cache_pool_size(&self) -> usize

Cardano transactions prover cache pool size

fn cardano_transactions_database_connection_pool_size(&self) -> usize

Cardano transactions database connection pool size

fn cardano_transactions_signing_config( &self, ) -> Option<CardanoTransactionsSigningConfig>

Cardano transactions signing configuration

fn cardano_blocks_transactions_signing_config( &self, ) -> Option<CardanoBlocksTransactionsSigningConfig>

Cardano blocks and transactions signing configuration

fn preload_security_parameter(&self) -> BlockNumber

Blocks offset, from the tip of the chain, to exclude during the cardano transactions preload

fn cardano_prover_max_hashes_allowed_by_request(&self) -> usize

Maximum number of hashes allowed by request to the Cardano prover.

This unified limit is applied to both transaction-hash and block-hash proving routes.

fn cardano_transactions_block_streamer_max_roll_forwards_per_poll( &self, ) -> usize

The maximum number of roll forwards during a poll of the block streamer when importing transactions.

fn cardano_transactions_block_streamer_throttling_interval(&self) -> Option<u64>

Minimum duration between two consecutive block streamer polls in milliseconds.

fn enable_metrics_server(&self) -> bool

Enable metrics server (Prometheus endpoint on /metrics).

fn metrics_server_ip(&self) -> String

Metrics HTTP Server IP.

fn metrics_server_port(&self) -> u16

Metrics HTTP Server listening port.

fn persist_usage_report_interval_in_seconds(&self) -> u64

Time interval at which usage metrics are persisted in event database (in seconds).

fn leader_aggregator_endpoint(&self) -> Option<String>

Leader aggregator endpoint

This is the endpoint of the aggregator that will be used to fetch the latest epoch settings and store the signer registrations when the aggregator is running in a follower mode. If this is not set, the aggregator will run in a leader mode.

fn custom_origin_tag_white_list(&self) -> Option<String>

Custom origin tag of client request added to the whitelist (comma separated list).

fn get_server_url(&self) -> StdResult<SanitizedUrlWithTrailingSlash>

Get the server URL.

fn get_network(&self) -> StdResult<CardanoNetwork>

Get a representation of the Cardano network.

fn get_dmq_network(&self) -> StdResult<DmqNetwork>

Get a representation of the DMQ network.

fn get_sqlite_dir(&self) -> PathBuf

Get the directory of the SQLite stores.

fn get_snapshot_dir(&self) -> StdResult<PathBuf>

Get the snapshots directory.

fn safe_epoch_retention_limit(&self) -> Option<u64>

Get the safe epoch retention limit.

fn compute_allowed_signed_entity_types_discriminants( &self, ) -> StdResult<BTreeSet<SignedEntityTypeDiscriminants>>

Compute the list of signed entity discriminants that are allowed to be processed.

fn allow_http_serve_directory(&self) -> bool

Check if the HTTP server can serve static directories.

fn get_leader_aggregator_epoch_settings_configuration( &self, ) -> StdResult<AggregatorEpochSettings>

leader aggregator only Infer the AggregatorEpochSettings from the configuration.

fn is_follower_aggregator(&self) -> bool

Check if the aggregator is running in follower mode.

fn compute_origin_tag_white_list(&self) -> HashSet<String>

White list for origin client request.

fn aggregate_signature_type(&self) -> AggregateSignatureType

Aggregate signature type

fn signature_processor_wait_delay_on_error_ms(&self) -> u64

Signature processor wait delay on error in milliseconds