Trait mithril_aggregator::MultiSigner

pub trait MultiSigner: Sync + Send {

    fn get_current_message<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<ProtocolMessage>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_current_message<'life0, 'async_trait>(
        &'life0 mut self, 
        message: ProtocolMessage
    ) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_current_beacon<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<Beacon>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_current_beacon<'life0, 'async_trait>(
        &'life0 mut self, 
        beacon: Beacon
    ) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_protocol_parameters<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolParameters>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_protocol_parameters<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        protocol_parameters: &'life1 ProtocolParameters
    ) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_next_protocol_parameters<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolParameters>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_stake_distribution<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<ProtocolStakeDistribution, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_next_stake_distribution<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<ProtocolStakeDistribution, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_stake_distribution<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        stakes: &'life1 ProtocolStakeDistribution
    ) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn compute_aggregate_verification_key<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        signers_with_stakes: &'life1 [SignerWithStake], 
        protocol_parameters: &'life2 ProtocolParameters
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolAggregateVerificationKey>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn register_signer<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        party_id: ProtocolPartyId, 
        verification_key: &'life1 ProtocolSignerVerificationKey
    ) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_signer<'life0, 'async_trait>(
        &'life0 self, 
        party_id: ProtocolPartyId
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolSignerVerificationKey>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_signers_with_stake<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SignerWithStake>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_next_signers_with_stake<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SignerWithStake>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn register_single_signature<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        signatures: &'life1 SingleSignatures
    ) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_multi_signature<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolMultiSignature>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn create_multi_signature<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolMultiSignature>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn create_certificate<'life0, 'async_trait>(
        &'life0 self, 
        beacon: Beacon, 
        previous_hash: String
    ) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn compute_stake_distribution_aggregate_verification_key<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn compute_next_stake_distribution_aggregate_verification_key<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn get_signers<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Signer>, ProtocolError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

}

MultiSigner is the cryptographic engine in charge of producing multi signatures from individual signatures

Required Methods

fn get_current_message<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Option<ProtocolMessage>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get current message

fn update_current_message<'life0, 'async_trait>(
&'life0 mut self,
message: ProtocolMessage
) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Update current message

fn get_current_beacon<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Option<Beacon>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get current beacon

fn update_current_beacon<'life0, 'async_trait>(
&'life0 mut self,
beacon: Beacon
) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Update current beacon

fn get_protocol_parameters<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolParameters>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get protocol parameters

fn update_protocol_parameters<'life0, 'life1, 'async_trait>(
&'life0 mut self,
protocol_parameters: &'life1 ProtocolParameters
) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Update protocol parameters

fn get_next_protocol_parameters<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolParameters>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get next protocol parameters

fn get_stake_distribution<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<ProtocolStakeDistribution, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get stake distribution

fn get_next_stake_distribution<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<ProtocolStakeDistribution, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get next stake distribution i.e. the stake distribution that will be used at the next epoch

fn update_stake_distribution<'life0, 'life1, 'async_trait>(
&'life0 mut self,
stakes: &'life1 ProtocolStakeDistribution
) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Update stake distribution

fn compute_aggregate_verification_key<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
signers_with_stakes: &'life1 [SignerWithStake],
protocol_parameters: &'life2 ProtocolParameters
) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolAggregateVerificationKey>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Compute aggregate verification key from stake distribution

fn register_signer<'life0, 'life1, 'async_trait>(
&'life0 mut self,
party_id: ProtocolPartyId,
verification_key: &'life1 ProtocolSignerVerificationKey
) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Register a signer

fn get_signer<'life0, 'async_trait>(
&'life0 self,
party_id: ProtocolPartyId
) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolSignerVerificationKey>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get signer

fn get_signers_with_stake<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<SignerWithStake>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get signers with stake

fn get_next_signers_with_stake<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<SignerWithStake>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get signers for the next epoch with their stake

fn register_single_signature<'life0, 'life1, 'async_trait>(
&'life0 mut self,
signatures: &'life1 SingleSignatures
) -> Pin<Box<dyn Future<Output = Result<(), ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Registers a single signature

fn get_multi_signature<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolMultiSignature>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Retrieves a multi signature from a message

fn create_multi_signature<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Option<ProtocolMultiSignature>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Creates a multi signature from single signatures

fn create_certificate<'life0, 'async_trait>(
&'life0 self,
beacon: Beacon,
previous_hash: String
) -> Pin<Box<dyn Future<Output = Result<Option<Certificate>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Creates a certificate from a multi signatures

Provided Methods

fn compute_stake_distribution_aggregate_verification_key<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<String>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Compute stake distribution aggregate verification key

fn compute_next_stake_distribution_aggregate_verification_key<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<String>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Compute next stake distribution aggregate verification key

fn get_signers<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<Signer>, ProtocolError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get signers